Misc cleanup, UI fixes and add PWA support.

+++ app/src/views/checklist/AllChecklist.vue
@@ -18,6 +18,7 @@ async function create() {
 }
 
 async function deleteChecklist(checklistID: ChecklistID) {
+  if (!confirm("Delete?")) return;
   loader.show();
   await checklistStore.delete(checklistID);
   loader.hide();
